// src/pages/smartReplenishment/pages/createrDispatchGoods/columns.d.ts
// 请确保此处与真实字段完全匹配
export enum keys {
  id = 'id',                        // 主键id
  barCode = 'barCode',              // 商品条码
  goodsCode = 'goodsCode',          // 商品编码
  goodsName = 'goodsName',          // 商品名称
  procurementDescr = 'procurementDescr',  // 采购分类说明
  procurementLv1 = 'procurementLv1',// 采购一级
  categroyLv1 = 'categroyLv1',      // 品类一级
  ptcnStatus = 'ptcnStatus',        // 采购分类状态
  retailPrice = 'retailPrice',      // 零售价
  stockQty = 'stockQty',            // 实时库存
  replenishmentCount = 'replenishmentCount'  // 补货数量
}